+++ CHANGELOG.md
@@ -26,6 +26,8 @@
 - Adds inter-component communication for map and bottom sheet view - ash
 - Fix createMockNetworkLayer and update tests - luc
 - Adds inter-component communication for map and bottom sheet view - ash
+- Adds SearchIcon from svg - Kieran
+- Styles SearchLink - Kieran
 - Fixes bug in filter bar when centering selected items - ash
 - Centers Fair titles in fairbooth preview - kieran
 - Adds some small styling adjustments to the hoursCollapsible - kieran